Adding an upload stage to the build firmware (#4)
Upload the .bin file from the build directory. The name is modified to
include the git hash.
The name is hardcoded in the `yml` file, to the project name. Ideally
the CMake should set the project name as an ENV variable.
Useful so that you can flash other PRs without building yourself.